backend/requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.34.2
sqlalchemy==2.0.20
pydantic==2.6.1
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.1
python-jose==3.3.0
bcrypt==4.0.1
httpx==0.28.1
python-dotenv==1.0.0
aiomysql
